+++ fastly/products/botmanagement/api_test.go
@@ -1,6 +1,7 @@
 package botmanagement_test
 
 import (
+	"context"
 	"testing"
 
 	"github.com/fastly/go-fastly/v13/fastly"
@@ -8,7 +9,19 @@ import (
 	"github.com/fastly/go-fastly/v13/fastly/products/botmanagement"
 	"github.com/fastly/go-fastly/v13/internal/productcore"
 	"github.com/fastly/go-fastly/v13/internal/test_utils"
+
+	"github.com/stretchr/testify/require"
 )
+
+var serviceID = fastly.TestDeliveryServiceID
+
+func TestUpdateConfigurationMissingContentGuard(t *testing.T) {
+	t.Parallel()
+
+	_, err := botmanagement.UpdateConfiguration(context.TODO(), nil, serviceID, botmanagement.ConfigureInput{ContentGuard: ""})
+
+	require.ErrorIs(t, err, botmanagement.ErrMissingContentGuard)
+}
 
 var functionalTests = []*test_utils.FunctionalTest{
 	productcore.NewDisableTest(&productcore.DisableTestInput{
@@ -31,6 +44,33 @@ var functionalTests = []*test_utils.FunctionalTest{
 		OpFn:      botmanagement.Get,
 		ProductID: botmanagement.ProductID,
 	}),
+	productcore.NewGetConfigurationTest(&productcore.GetConfigurationTestInput[botmanagement.ConfigureOutput]{
+		Phase:     "default",
+		OpFn:      botmanagement.GetConfiguration,
+		ProductID: botmanagement.ProductID,
+		CheckOutputFn: func(t *testing.T, tc *test_utils.FunctionalTest, o botmanagement.ConfigureOutput) {
+			require.NotNilf(t, o.Configuration.ContentGuard, "test '%s'", tc.Name)
+			require.Equalf(t, "off", *o.Configuration.ContentGuard, "test '%s'", tc.Name)
+		},
+	}),
+	productcore.NewUpdateConfigurationTest(&productcore.UpdateConfigurationTestInput[botmanagement.ConfigureOutput, botmanagement.ConfigureInput]{
+		OpFn:      botmanagement.UpdateConfiguration,
+		Input:     botmanagement.ConfigureInput{ContentGuard: "on"},
+		ProductID: botmanagement.ProductID,
+		CheckOutputFn: func(t *testing.T, tc *test_utils.FunctionalTest, o botmanagement.ConfigureOutput) {
+			require.NotNilf(t, o.Configuration.ContentGuard, "test '%s'", tc.Name)
+			require.Equalf(t, "on", *o.Configuration.ContentGuard, "test '%s'", tc.Name)
+		},
+	}),
+	productcore.NewGetConfigurationTest(&productcore.GetConfigurationTestInput[botmanagement.ConfigureOutput]{
+		Phase:     "after update",
+		OpFn:      botmanagement.GetConfiguration,
+		ProductID: botmanagement.ProductID,
+		CheckOutputFn: func(t *testing.T, tc *test_utils.FunctionalTest, o botmanagement.ConfigureOutput) {
+			require.NotNilf(t, o.Configuration.ContentGuard, "test '%s'", tc.Name)
+			require.Equalf(t, "on", *o.Configuration.ContentGuard, "test '%s'", tc.Name)
+		},
+	}),
 	productcore.NewDisableTest(&productcore.DisableTestInput{
 		OpFn: botmanagement.Disable,
 	}),
@@ -42,6 +82,6 @@ var functionalTests = []*test_utils.FunctionalTest{
 	}),
 }
 
-func TestEnablementDelivery(t *testing.T) {
+func TestEnablementAndConfigurationDelivery(t *testing.T) {
 	test_utils.ExecuteFunctionalTests(t, functionalTests, fastly.TestDeliveryServiceID)
 }